Split body profile from account settings
