+++ frontend/src/pages/Profile.tsx
@@ -1,12 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import {
-    Alert,
     Box,
-    Button,
-    Dialog,
-    DialogActions,
-    DialogContent,
-    DialogTitle,
     FormControl,
     Link,
     InputLabel,
@@ -20,8 +14,6 @@ import { useTheme } from '@mui/material/styles';
 import { Link as RouterLink } from 'react-router-dom';
 import CalorieTargetBanner from '../components/CalorieTargetBanner';
 import { activityLevelOptions } from '../constants/activityLevels';
-import ProfilePhotoCard from '../components/ProfilePhotoCard';
-import { useTransientStatus } from '../hooks/useTransientStatus';
 import type { UserProfilePatchPayload } from '../context/authContext';
 import { useAuth } from '../context/useAuth';
 import { useUserProfileQuery } from '../queries/userProfile';
@@ -29,11 +21,9 @@ import AppPage from '../ui/AppPage';
 import AppCard from '../ui/AppCard';
 import InlineStatusLine from '../ui/InlineStatusLine';
 import SectionHeader from '../ui/SectionHeader';
-import { getApiErrorMessage } from '../utils/apiError';
 import { getDefaultHeightUnitForWeightUnit } from '../utils/unitPreferences';
 
 const AUTOSAVE_DELAY_MS = 450;
-const MIN_PASSWORD_LENGTH = 8;
 
 type ParsedHeight = {
     cm: number;
@@ -82,20 +72,13 @@ function buildFeetInchesHeightPatch(feet: string, inches: string): UserProfilePa
 }
 
 /**
- * Profile is the dedicated page for editing user-specific profile fields used for calorie math.
+ * Profile is focused on body-profile inputs used for calorie math (BMR/TDEE).
+ * Account settings (photo/password) live in Settings.
  */
 const Profile: React.FC = () => {
     const theme = useTheme();
-    const { user, updateProfile, changePassword } = useAuth();
+    const { user, updateProfile } = useAuth();
     const sectionGap = theme.custom.layout.page.sectionGap;
-    const { status: accountStatus, showStatus: showAccountStatus, clearStatus: clearAccountStatus } = useTransientStatus();
-    const [passwordError, setPasswordError] = useState('');
-    const [isChangingPassword, setIsChangingPassword] = useState(false);
-    const [isPasswordDialogOpen, setIsPasswordDialogOpen] = useState(false);
-
-    const [currentPassword, setCurrentPassword] = useState('');
-    const [newPassword, setNewPassword] = useState('');
-    const [confirmPassword, setConfirmPassword] = useState('');
 
     const autosaveTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
     const savedMessageTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
@@ -247,100 +230,18 @@ const Profile: React.FC = () => {
         [flushAutosave]
     );
 
-    /**
-     * Open the change-password dialog and clear any prior error state.
-     */
-    const handlePasswordDialogOpen = () => {
-        clearAccountStatus();
-        resetPasswordDialogFields();
-        setIsPasswordDialogOpen(true);
-    };
-
-    /**
-     * Clear sensitive input values used by the password dialog.
-     */
-    const resetPasswordDialogFields = () => {
-        setPasswordError('');
-        setCurrentPassword('');
-        setNewPassword('');
-        setConfirmPassword('');
-    };
-
-    /**
-     * Close the change-password dialog and clear sensitive input values.
-     */
-    const closePasswordDialog = () => {
-        setIsPasswordDialogOpen(false);
-        resetPasswordDialogFields();
-    };
-
-    /**
-     * Close the change-password dialog (unless a request is in-flight).
-     */
-    const handlePasswordDialogClose = () => {
-        if (isChangingPassword) return;
-        closePasswordDialog();
-    };
-
-    /**
-     * Change the current user's password after validating basic client-side constraints.
-     */
-    const handlePasswordChange = async () => {
-        clearAccountStatus();
-        setPasswordError('');
-
-        if (!currentPassword) {
-            setPasswordError('Please enter your current password.');
-            return;
-        }
-
-        if (newPassword.length < MIN_PASSWORD_LENGTH) {
-            setPasswordError(`New password must be at least ${MIN_PASSWORD_LENGTH} characters.`);
-            return;
-        }
-
-        if (newPassword !== confirmPassword) {
-            setPasswordError('New passwords do not match.');
-            return;
-        }
-
-        if (currentPassword === newPassword) {
-            setPasswordError('New password must be different from your current password.');
-            return;
-        }
-
-        setIsChangingPassword(true);
-        try {
-            await changePassword(currentPassword, newPassword);
-            showAccountStatus('Password updated.', 'success');
-            closePasswordDialog();
-        } catch (err) {
-            setPasswordError(getApiErrorMessage(err) ?? 'Failed to update password.');
-        } finally {
-            setIsChangingPassword(false);
-        }
-    };
-
-    /**
-     * Keep the password form accessible by handling Enter-to-submit and preventing full-page reloads.
-     */
-    const handlePasswordSubmit = (e: React.FormEvent) => {
-        e.preventDefault();
-        void handlePasswordChange();
-    };
-
     return (
         <AppPage maxWidth="content">
             <Stack spacing={sectionGap} useFlexGap>
                 <CalorieTargetBanner />
 
-                <Typography color="text.secondary">
-                    Changes save automatically. Update the inputs below to recalculate your calorie target (TDEE +/- goal deficit).
-                </Typography>
-
-                <ProfilePhotoCard description="Used for your avatar in the app bar." />
-
                 <AppCard>
+                    <SectionHeader
+                        title="Body profile"
+                        subtitle="Changes save automatically. Update these inputs to recalculate your estimated baseline burn (TDEE)."
+                        sx={{ mb: 0.5 }}
+                    />
+
                     <InlineStatusLine status={autosaveStatusLine} sx={{ mb: 1 }} ariaLive="off" />
 
                     <Stack spacing={2}>
@@ -439,108 +340,6 @@ const Profile: React.FC = () => {
                         </FormControl>
                     </Stack>
                 </AppCard>
-
-                <AppCard>
-                    <SectionHeader
-                        title="Account"
-                        subtitle="View your email address and update your password."
-                        actions={
-                            <Button variant="outlined" onClick={handlePasswordDialogOpen}>
-                                Change Password
-                            </Button>
-                        }
-                        sx={{ mb: 0.5 }}
-                    />
-
-                    <InlineStatusLine status={accountStatus} sx={{ mb: 1 }} />
-
-                    <Stack spacing={1.5}>
-                        <Typography variant="body2" color="text.secondary">
-                            Email
-                        </Typography>
-                        <Box
-                            sx={{
-                                px: 2,
-                                py: 1.5,
-                                border: '1px solid',
-                                borderColor: 'divider',
-                                borderRadius: 1,
-                                backgroundColor: 'action.hover'
-                            }}
-                        >
-                            <Typography sx={{ wordBreak: 'break-word' }}>{user?.email ?? ''}</Typography>
-                        </Box>
-                    </Stack>
-                </AppCard>
-
-                <Dialog
-                    open={isPasswordDialogOpen}
-                    onClose={handlePasswordDialogClose}
-                    fullWidth
-                    maxWidth="xs"
-                >
-                    <DialogTitle>Change password</DialogTitle>
-                    <DialogContent>
-                        <Stack
-                            spacing={2}
-                            component="form"
-                            id="change-password-form"
-                            onSubmit={handlePasswordSubmit}
-                            sx={{ pt: 1 }}
-                        >
-                            {passwordError && <Alert severity="error">{passwordError}</Alert>}
-
-                            <TextField
-                                label="Current Password"
-                                type="password"
-                                autoComplete="current-password"
-                                value={currentPassword}
-                                onChange={(e) => setCurrentPassword(e.target.value)}
-                                disabled={isChangingPassword}
-                                required
-                                fullWidth
-                            />
-
-                            <TextField
-                                label="New Password"
-                                type="password"
-                                autoComplete="new-password"
-                                value={newPassword}
-                                onChange={(e) => setNewPassword(e.target.value)}
-                                helperText={`At least ${MIN_PASSWORD_LENGTH} characters.`}
-                                disabled={isChangingPassword}
-                                inputProps={{ minLength: MIN_PASSWORD_LENGTH }}
-                                required
-                                fullWidth
-                            />
-
-                            <TextField
-                                label="Confirm New Password"
-                                type="password"
-                                autoComplete="new-password"
-                                value={confirmPassword}
-                                onChange={(e) => setConfirmPassword(e.target.value)}
-                                disabled={isChangingPassword}
-                                inputProps={{ minLength: MIN_PASSWORD_LENGTH }}
-                                required
-                                fullWidth
-                            />
-                        </Stack>
-                    </DialogContent>
-                    <DialogActions sx={{ px: 3, pb: 2 }}>
-                        <Button onClick={handlePasswordDialogClose} disabled={isChangingPassword}>
-                            Cancel
-                        </Button>
-                        <Button
-                            type="submit"
-                            form="change-password-form"
-                            variant="contained"
-                            disabled={isChangingPassword}
-                        >
-                            {isChangingPassword ? 'Updating...' : 'Update Password'}
-                        </Button>
-                    </DialogActions>
-                </Dialog>
             </Stack>
         </AppPage>
     );
